re-refactor the pct safe to ordinal grade calculations; add v1 benchmark grade looking at individual test results, rather than roll-ups of hazard scores
